# app/api/pagamento_grp_descontos/route.ts
import { NextRequest, NextResponse } from "next/server";
import { prisma } from "@/lib/prisma";
import { pagamentoGrpDescontoSchema } from "@/lib/validation/pagamentoGrpDesconto";

// GET /api/pagamento_grp_descontos — paginated list
export async function GET(req: NextRequest) {
  const { searchParams } = new URL(req.url);
  const page = Math.max(1, Number(searchParams.get("page") ?? "1"));
  const limit = Math.min(100, Math.max(1, Number(searchParams.get("limit") ?? "20")));
  const [data, total] = await Promise.all([
    prisma.pagamentoGrpDesconto.findMany({ skip: (page - 1) * limit, take: limit }),
    prisma.pagamentoGrpDesconto.count(),
  ]);
  return NextResponse.json({ data, total, page, limit });
}

// POST /api/pagamento_grp_descontos — create
export async function POST(req: NextRequest) {
  const parsed = pagamentoGrpDescontoSchema.safeParse(await req.json());
  if (!parsed.success) {
    return NextResponse.json({ errors: parsed.error.flatten() }, { status: 400 });
  }
  const created = await prisma.pagamentoGrpDesconto.create({ data: parsed.data });
  return NextResponse.json(created, { status: 201 });
}