# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Generated Application API"
  version: "1.0.0"
  description: |
    API generated by RNC from legacy source code.
    Business rules are preserved from the original system.
servers:
  - url: http://localhost:3059/api
    description: Local development
paths:
  /auditorias:
    get:
      summary: List Auditorias
      operationId: listAuditorias
      tags: [Auditoria]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of Auditorias
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuditoriaList"}
    post:
      summary: Create Auditoria
      operationId: createAuditoria
      tags: [Auditoria]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AuditoriaInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Auditoria"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/auditorias/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get Auditoria by id
      operationId: getAuditoriaById
      tags: [Auditoria]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Auditoria"}
        "404": {description: Not found}
    put:
      summary: Update Auditoria
      operationId: updateAuditoria
      tags: [Auditoria]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AuditoriaInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Auditoria"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete Auditoria
      operationId: deleteAuditoria
      tags: [Auditoria]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
  /pagamentos:
    get:
      summary: List Pagamentos
      operationId: listPagamentos
      tags: [Pagamento]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of Pagamentos
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PagamentoList"}
    post:
      summary: Create Pagamento
      operationId: createPagamento
      tags: [Pagamento]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PagamentoInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Pagamento"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/pagamentos/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get Pagamento by id
      operationId: getPagamentoById
      tags: [Pagamento]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Pagamento"}
        "404": {description: Not found}
    put:
      summary: Update Pagamento
      operationId: updatePagamento
      tags: [Pagamento]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PagamentoInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Pagamento"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete Pagamento
      operationId: deletePagamento
      tags: [Pagamento]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
  /pagamento_grp_descontos:
    get:
      summary: List PagamentoGrpDescontos
      operationId: listPagamentoGrpDescontos
      tags: [PagamentoGrpDesconto]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of PagamentoGrpDescontos
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PagamentoGrpDescontoList"}
    post:
      summary: Create PagamentoGrpDesconto
      operationId: createPagamentoGrpDesconto
      tags: [PagamentoGrpDesconto]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PagamentoGrpDescontoInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PagamentoGrpDesconto"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/pagamento_grp_descontos/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get PagamentoGrpDesconto by id
      operationId: getPagamentoGrpDescontoById
      tags: [PagamentoGrpDesconto]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PagamentoGrpDesconto"}
        "404": {description: Not found}
    put:
      summary: Update PagamentoGrpDesconto
      operationId: updatePagamentoGrpDesconto
      tags: [PagamentoGrpDesconto]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PagamentoGrpDescontoInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PagamentoGrpDesconto"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete PagamentoGrpDesconto
      operationId: deletePagamentoGrpDesconto
      tags: [PagamentoGrpDesconto]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
...
  /beneficiarios:
    get:
      summary: List Beneficiarios
      operationId: listBeneficiarios
      tags: [Beneficiario]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of Beneficiarios
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeneficiarioList"}
    post:
      summary: Create Beneficiario
      operationId: createBeneficiario
      tags: [Beneficiario]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeneficiarioInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Beneficiario"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/beneficiarios/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get Beneficiario by id
      operationId: getBeneficiarioById
      tags: [Beneficiario]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Beneficiario"}
        "404": {description: Not found}
    put:
      summary: Update Beneficiario
      operationId: updateBeneficiario
      tags: [Beneficiario]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeneficiarioInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Beneficiario"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete Beneficiario
      operationId: deleteBeneficiario
      tags: [Beneficiario]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
  /beneficiario_grp_dependentes:
    get:
      summary: List BeneficiarioGrpDependentes
      operationId: listBeneficiarioGrpDependentes
      tags: [BeneficiarioGrpDependente]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of BeneficiarioGrpDependentes
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeneficiarioGrpDependenteList"}
    post:
      summary: Create BeneficiarioGrpDependente
      operationId: createBeneficiarioGrpDependente
      tags: [BeneficiarioGrpDependente]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeneficiarioGrpDependenteInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeneficiarioGrpDependente"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/beneficiario_grp_dependentes/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get BeneficiarioGrpDependente by id
      operationId: getBeneficiarioGrpDependenteById
      tags: [BeneficiarioGrpDependente]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeneficiarioGrpDependente"}
        "404": {description: Not found}
    put:
      summary: Update BeneficiarioGrpDependente
      operationId: updateBeneficiarioGrpDependente
      tags: [BeneficiarioGrpDependente]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeneficiarioGrpDependenteInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeneficiarioGrpDependente"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete BeneficiarioGrpDependente
      operationId: deleteBeneficiarioGrpDependente
      tags: [BeneficiarioGrpDependente]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
components:
  schemas:
    Auditoria:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        numAuditoria:
          type: string
          nullable: true
          description: "Num Auditoria"
        dtEvento:
          type: string
          nullable: true
          description: "Dt Evento"
        hrEvento:
          type: string
          nullable: true
          description: "Hr Evento"
        tsEvento:
          type: string
          nullable: true
          description: "Ts Evento"
        codAcao:
          type: string
          nullable: true
          description: "Cod Acao"
        codModulo:
          type: string
          nullable: true
          description: "Cod Modulo"
        desAcao:
          type: string
          nullable: true
          description: "Des Acao"
        tipoEntidade:
          type: string
          nullable: true
          description: "Tipo Entidade"
        idEntidade:
          type: string
          nullable: true
          description: "Id Entidade"
        numCpfAfetado:
          type: string
          nullable: true
          description: "Num Cpf Afetado"
        usrEvento:
          type: string
          nullable: true
          description: "Usr Evento"
        nomeUsuario:
          type: string
          nullable: true
          description: "Nome Usuario"
        codPerfil:
          type: string
          nullable: true
          description: "Cod Perfil"
        codLotacao:
          type: string
          nullable: true
          description: "Cod Lotacao"
        ipOrigem:
          type: string
          nullable: true
          description: "Ip Origem"
        idSessao:
          type: string
          nullable: true
          description: "Id Sessao"
        numCicloBatch:
          type: string
          nullable: true
          description: "Num Ciclo Batch"
        numSeqBatch:
          type: string
          nullable: true
          description: "Num Seq Batch"
        nomJobBatch:
          type: string
          nullable: true
          description: "Nom Job Batch"
        sitBatch:
          type: string
          nullable: true
          description: "Sit Batch"
        desErroBatch:
          type: string
          nullable: true
          description: "Des Erro Batch"
        idCorrelacao:
          type: string
          nullable: true
          description: "Id Correlacao"
        numSeqCorrelacao:
          type: string
          nullable: true
          description: "Num Seq Correlacao"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    AuditoriaInput:
      type: object
      properties:
        numAuditoria:
          type: string
          nullable: true
          description: "Num Auditoria"
        dtEvento:
          type: string
          nullable: true
          description: "Dt Evento"
        hrEvento:
          type: string
          nullable: true
          description: "Hr Evento"
        tsEvento:
          type: string
          nullable: true
          description: "Ts Evento"
        codAcao:
          type: string
          nullable: true
          description: "Cod Acao"
        codModulo:
          type: string
          nullable: true
          description: "Cod Modulo"
        desAcao:
          type: string
          nullable: true
          description: "Des Acao"
        tipoEntidade:
          type: string
          nullable: true
          description: "Tipo Entidade"
        idEntidade:
          type: string
          nullable: true
          description: "Id Entidade"
        numCpfAfetado:
          type: string
          nullable: true
          description: "Num Cpf Afetado"
        usrEvento:
          type: string
          nullable: true
          description: "Usr Evento"
        nomeUsuario:
          type: string
          nullable: true
          description: "Nome Usuario"
        codPerfil:
          type: string
          nullable: true
          description: "Cod Perfil"
        codLotacao:
          type: string
          nullable: true
          description: "Cod Lotacao"
        ipOrigem:
          type: string
          nullable: true
          description: "Ip Origem"
        idSessao:
          type: string
          nullable: true
          description: "Id Sessao"
        numCicloBatch:
          type: string
          nullable: true
          description: "Num Ciclo Batch"
        numSeqBatch:
          type: string
          nullable: true
          description: "Num Seq Batch"
        nomJobBatch:
          type: string
          nullable: true
          description: "Nom Job Batch"
        sitBatch:
          type: string
          nullable: true
          description: "Sit Batch"
        desErroBatch:
          type: string
          nullable: true
          description: "Des Erro Batch"
        idCorrelacao:
          type: string
          nullable: true
          description: "Id Correlacao"
        numSeqCorrelacao:
          type: string
          nullable: true
          description: "Num Seq Correlacao"
    AuditoriaList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Auditoria"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
...
    PagamentoInput:
      type: object
      properties:
        numPagamento:
          type: string
          nullable: true
          description: "Num Pagamento"
        numCpf:
          type: string
          nullable: true
          description: "Num Cpf"
        numInscricao:
          type: string
          nullable: true
          description: "Num Inscricao"
        codPrograma:
          type: string
          nullable: true
          description: "Cod Programa"
        anoMesRef:
          type: string
          nullable: true
          description: "Ano Mes Ref"
        numCiclo:
          type: string
          nullable: true
          description: "Num Ciclo"
        vlrBruto:
          type: string
          nullable: true
          description: "Vlr Bruto"
        vlrLiquido:
          type: string
          nullable: true
          description: "Vlr Liquido"
        vlrDescontoTotal:
          type: string
          nullable: true
          description: "Vlr Desconto Total"
        sitPagamento:
          type: string
          nullable: true
          description: "Sit Pagamento"
        dtGeracao:
          type: string
          nullable: true
          description: "Dt Geracao"
        hrGeracao:
          type: string
          nullable: true
          description: "Hr Geracao"
        dtEmissao:
          type: string
          nullable: true
          description: "Dt Emissao"
        dtConfirmacao:
          type: string
          nullable: true
          description: "Dt Confirmacao"
        dtCancelamento:
          type: string
          nullable: true
          description: "Dt Cancelamento"
        motCancelamento:
          type: string
          nullable: true
          description: "Mot Cancelamento"
        codBanco:
          type: string
          nullable: true
          description: "Cod Banco"
        codAgencia:
          type: string
          nullable: true
          description: "Cod Agencia"
        numConta:
          type: string
          nullable: true
          description: "Num Conta"
        tipoConta:
          type: string
          nullable: true
          description: "Tipo Conta"
        codOperacao:
          type: string
          nullable: true
          description: "Cod Operacao"
        numObSiafi:
          type: string
          nullable: true
          description: "Num Ob Siafi"
        numNeSiafi:
          type: string
          nullable: true
          description: "Num Ne Siafi"
        codUgEmitente:
          type: string
          nullable: true
          description: "Cod Ug Emitente"
        codGestao:
          type: string
          nullable: true
          description: "Cod Gestao"
        sitIntegSiafi:
          type: string
          nullable: true
          description: "Sit Integ Siafi"
        dtConciliacao:
          type: string
          nullable: true
          description: "Dt Conciliacao"
        sitConciliacao:
          type: string
          nullable: true
          description: "Sit Conciliacao"
        vlrConciliado:
          type: string
          nullable: true
          description: "Vlr Conciliado"
        codRetornoBanco:
          type: string
          nullable: true
          description: "Cod Retorno Banco"
        desRetornoBanco:
          type: string
          nullable: true
          description: "Des Retorno Banco"
        hashArqRemessa:
          type: string
          nullable: true
          description: "Hash Arq Remessa"
        hashArqRetorno:
          type: string
          nullable: true
          description: "Hash Arq Retorno"
        dtInclusao:
          type: string
          nullable: true
          description: "Dt Inclusao"
        hrInclusao:
          type: string
          nullable: true
          description: "Hr Inclusao"
        usrInclusao:
          type: string
          nullable: true
          description: "Usr Inclusao"
        dtUltAlteracao:
          type: string
          nullable: true
          description: "Dt Ult Alteracao"
        hrUltAlteracao:
          type: string
          nullable: true
          description: "Hr Ult Alteracao"
        usrUltAlteracao:
          type: string
          nullable: true
          description: "Usr Ult Alteracao"
    PagamentoList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Pagamento"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
    PagamentoGrpDesconto:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        tipoDesconto:
          type: string
          nullable: true
          description: "Tipo Desconto"
        vlrDesconto:
          type: string
          nullable: true
          description: "Vlr Desconto"
        pctDesconto:
          type: string
          nullable: true
          description: "Pct Desconto"
        numProcesso:
          type: string
          nullable: true
          description: "Num Processo"
        dtInicioDsct:
          type: string
          nullable: true
          description: "Dt Inicio Dsct"
        dtFimDsct:
          type: string
          nullable: true
          description: "Dt Fim Dsct"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    PagamentoGrpDescontoInput:
      type: object
      properties:
        tipoDesconto:
          type: string
          nullable: true
          description: "Tipo Desconto"
        vlrDesconto:
          type: string
          nullable: true
          description: "Vlr Desconto"
        pctDesconto:
          type: string
          nullable: true
          description: "Pct Desconto"
        numProcesso:
          type: string
          nullable: true
          description: "Num Processo"
        dtInicioDsct:
          type: string
          nullable: true
          description: "Dt Inicio Dsct"
        dtFimDsct:
          type: string
          nullable: true
          description: "Dt Fim Dsct"
    PagamentoGrpDescontoList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/PagamentoGrpDesconto"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
...